fix total price in db

+++ app/controllers/orders_controller.rb
@@ -1,6 +1,11 @@
 class OrdersController < ApplicationController
   def index
-    @orders = current_user.orders
+    if current_user
+      @orders = current_user.orders
+    else
+      flash[:non_user] = "You must be logged in to view your orders"
+      render file: 'public/404'
+    end
   end
 
   def show
@@ -8,7 +13,10 @@ class OrdersController < ApplicationController
   end
 
   def create
-    if current_user
+    if session[:cart].nil? || session[:cart] == {}
+      flash[:no_items] = "There's nothing in your cart."
+      redirect_to items_path
+    elsif current_user && session[:cart]
       current_user.set_order(session[:cart])
       session.delete(:cart)
       flash[:success] = 'Order was successfully placed'
